Cambios menores
- se crearon multiples clases
- se cambiaron los parametros de punto a Int

+++ src/puntos/Interseccion.scala
@@ -1,5 +1,5 @@
 package puntos
 
-class Interseccion(x: Double,y: Double, nombre : String = "") extends Punto(x,y){
+class Interseccion(x: Int,y: Int, nombre : String = "") extends Punto(x,y){
   
 }
